+++ src/lib/index.ts
@@ -1,6 +1,9 @@
 export { clickOutside } from '$lib/actions/click-outside.js'
 
-export { default as Connection } from '$lib/components/Connection.svelte'
+export {
+  default as Connection,
+  type ConnectionStatus
+} from '$lib/components/Connection.svelte'
 export { default as ToggleButton } from '$lib/components/ToggleButton.svelte'
 
 export { Error } from '$lib/error.js'